requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
gunicorn==23.0.0
packaging==24.2
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
